tests: robot: abstract simulator class and deduplicate code
Create a base class for simulator and reduce code duplication in
qemu/native sim classes.
