+++ tests/robot/NativeSimLibrary.py
@@ -3,23 +3,19 @@
 """Robot Framework library for interacting with Zephyr native_sim binaries via stdin/stdout."""
 
 import atexit
-import re
 import subprocess
 import threading
-import time
+
+from SimLibraryBase import SimLibraryBase
 
 
-class NativeSimLibrary:
+class NativeSimLibrary(SimLibraryBase):
     """Keyword library that starts a native_sim binary and provides UART-like interaction
     over its stdin/stdout (requires CONFIG_UART_NATIVE_PTY_0_ON_STDINOUT=y)."""
 
-    ROBOT_LIBRARY_SCOPE = 'SUITE'
-
     def __init__(self):
+        super().__init__()
         self._process = None
-        self._buffer = ''
-        self._lock = threading.Lock()
-        self._reader = None
 
     def start_native_binary(self, binary):
         """Start the native_sim binary and begin capturing its output."""
@@ -39,12 +35,8 @@ class NativeSimLibrary:
             data = self._process.stdout.read(256)
             if not data:
                 break
-            # Strip ANSI escape sequences before buffering
-            text = re.sub(
-                r'\x1b\[[0-9;]*[mABCDEFGHJKSTfhilmnsu]', '', data.decode('utf-8', errors='replace')
-            )
             with self._lock:
-                self._buffer += text
+                self._buffer += self._strip_ansi(data)
 
     def write_to_uart(self, text):
         """Write a line to the binary's stdin, simulating UART input."""
@@ -52,36 +44,6 @@ class NativeSimLibrary:
             raise RuntimeError('Native binary is not running')
         self._process.stdin.write((text + '\n').encode())
         self._process.stdin.flush()
-
-    def wait_for_output(self, pattern, timeout=15):
-        """Wait until *pattern* (regex) appears in the binary's output."""
-        end_time = time.time() + float(timeout)
-        regex = re.compile(pattern, re.MULTILINE)
-        while time.time() < end_time:
-            with self._lock:
-                m = regex.search(self._buffer)
-                if m:
-                    return m.group(0)
-            time.sleep(0.05)
-        with self._lock:
-            captured = self._buffer
-        raise AssertionError(
-            f"Pattern '{pattern}' not found within {timeout}s.\nCaptured output:\n{captured}"
-        )
-
-    def wait_for_literal_output(self, text, timeout=15):
-        """Wait until the literal *text* string appears in the binary's output."""
-        end_time = time.time() + float(timeout)
-        while time.time() < end_time:
-            with self._lock:
-                if text in self._buffer:
-                    return text
-            time.sleep(0.05)
-        with self._lock:
-            captured = self._buffer
-        raise AssertionError(
-            f"Text '{text}' not found within {timeout}s.\nCaptured output:\n{captured}"
-        )
 
     def stop_native_binary(self):
         """Terminate the native binary (called on suite teardown)."""
